Make needed changes to feature management infrastructure
