+++ common/src/main/java/com/wynntils/commands/FeatureCommand.java
@@ -8,14 +8,10 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.suggestion.SuggestionProvider;
-import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.commands.Command;
 import com.wynntils.core.components.Managers;
 import com.wynntils.core.config.Category;
-import com.wynntils.core.features.DebugFeature;
 import com.wynntils.core.features.Feature;
-import com.wynntils.core.features.StateManagedFeature;
-import com.wynntils.core.features.UserFeature;
 import java.util.List;
 import java.util.Optional;
 import net.minecraft.ChatFormatting;
@@ -23,16 +19,12 @@ import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
 import net.minecraft.commands.SharedSuggestionProvider;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.HoverEvent;
 import net.minecraft.network.chat.MutableComponent;
 
 public class FeatureCommand extends Command {
-    private static final SuggestionProvider<CommandSourceStack> USER_FEATURE_SUGGESTION_PROVIDER =
+    private static final SuggestionProvider<CommandSourceStack> FEATURE_SUGGESTION_PROVIDER =
             (context, builder) -> SharedSuggestionProvider.suggest(
-                    Managers.Feature.getFeatures().stream()
-                            .filter(feature -> feature instanceof UserFeature && isVisible(feature))
-                            .map(Feature::getShortName),
-                    builder);
+                    Managers.Feature.getFeatures().stream().map(Feature::getShortName), builder);
 
     @Override
     public String getCommandName() {
@@ -50,27 +42,21 @@ public class FeatureCommand extends Command {
                 .then(Commands.literal("list").executes(this::listFeatures))
                 .then(Commands.literal("enable")
                         .then(Commands.argument("feature", StringArgumentType.word())
-                                .suggests(USER_FEATURE_SUGGESTION_PROVIDER)
+                                .suggests(FEATURE_SUGGESTION_PROVIDER)
                                 .executes(this::enableFeature)))
                 .then(Commands.literal("disable")
                         .then(Commands.argument("feature", StringArgumentType.word())
-                                .suggests(USER_FEATURE_SUGGESTION_PROVIDER)
+                                .suggests(FEATURE_SUGGESTION_PROVIDER)
                                 .executes(this::disableFeature)))
                 .then(Commands.literal("reload")
                         .then(Commands.argument("feature", StringArgumentType.word())
-                                .suggests(USER_FEATURE_SUGGESTION_PROVIDER)
+                                .suggests(FEATURE_SUGGESTION_PROVIDER)
                                 .executes(this::reloadFeature)))
                 .executes(this::syntaxError);
     }
 
-    private static boolean isVisible(Feature feature) {
-        if (!(feature instanceof DebugFeature)) return true;
-        return WynntilsMod.isDevelopmentEnvironment();
-    }
-
     private int listFeatures(CommandContext<CommandSourceStack> context) {
         List<Feature> features = Managers.Feature.getFeatures().stream()
-                .filter(FeatureCommand::isVisible)
                 .sorted(Feature::compareTo)
                 .toList();
 
@@ -80,29 +66,8 @@ public class FeatureCommand extends Command {
         Category lastCategory = null;
 
         for (Feature feature : features) {
-            Class<?> superclass = feature.getClass().getSuperclass();
-
-            ChatFormatting color;
             String translatedName = feature.getTranslatedName();
-
-            if (feature instanceof DebugFeature) {
-                color = ChatFormatting.YELLOW;
-
-                if (feature.isEnabled()) {
-                    translatedName += " {ENABLED DEBUG}";
-                } else {
-                    translatedName += " {DISABLED DEBUG}";
-                }
-            } else {
-                if (feature.isEnabled()) {
-                    color = ChatFormatting.GREEN;
-                } else {
-                    color = ChatFormatting.RED;
-                }
-                if (feature instanceof StateManagedFeature) {
-                    translatedName += " {SYSTEM CONTROLLED}";
-                }
-            }
+            ChatFormatting color = feature.isEnabled() ? ChatFormatting.GREEN : ChatFormatting.RED;
 
             if (lastCategory != feature.getCategory()) {
                 lastCategory = feature.getCategory();
@@ -112,10 +77,7 @@ public class FeatureCommand extends Command {
             }
 
             response.append(Component.literal("\n - ").withStyle(ChatFormatting.GRAY))
-                    .append(Component.literal(translatedName)
-                            .withStyle(style -> style.withHoverEvent(new HoverEvent(
-                                    HoverEvent.Action.SHOW_TEXT, Component.literal(superclass.getSimpleName()))))
-                            .withStyle(color));
+                    .append(Component.literal(translatedName).withStyle(color));
         }
 
         context.getSource().sendSuccess(response, false);
@@ -128,11 +90,12 @@ public class FeatureCommand extends Command {
 
         Optional<Feature> featureOptional = Managers.Feature.getFeatureFromString(featureName);
 
-        if (featureOptional.isEmpty() || !(featureOptional.get() instanceof UserFeature feature)) {
+        if (featureOptional.isEmpty()) {
             context.getSource()
                     .sendFailure(Component.literal("Feature not found!").withStyle(ChatFormatting.RED));
             return 0;
         }
+        Feature feature = featureOptional.get();
 
         if (feature.isEnabled()) {
             context.getSource()
@@ -142,7 +105,6 @@ public class FeatureCommand extends Command {
         }
 
         feature.setUserEnabled(true);
-        feature.tryUserToggle();
 
         if (!feature.isEnabled()) {
             context.getSource()
@@ -167,11 +129,13 @@ public class FeatureCommand extends Command {
 
         Optional<Feature> featureOptional = Managers.Feature.getFeatureFromString(featureName);
 
-        if (featureOptional.isEmpty() || !(featureOptional.get() instanceof UserFeature feature)) {
+        if (featureOptional.isEmpty()) {
             context.getSource()
                     .sendFailure(Component.literal("Feature not found!").withStyle(ChatFormatting.RED));
             return 0;
         }
+
+        Feature feature = featureOptional.get();
 
         if (!feature.isEnabled()) {
             context.getSource()
@@ -181,7 +145,6 @@ public class FeatureCommand extends Command {
         }
 
         feature.setUserEnabled(false);
-        feature.tryUserToggle();
 
         if (feature.isEnabled()) {
             context.getSource()
@@ -206,11 +169,13 @@ public class FeatureCommand extends Command {
 
         Optional<Feature> featureOptional = Managers.Feature.getFeatureFromString(featureName);
 
-        if (featureOptional.isEmpty() || !(featureOptional.get() instanceof UserFeature feature)) {
+        if (featureOptional.isEmpty()) {
             context.getSource()
                     .sendFailure(Component.literal("Feature not found!").withStyle(ChatFormatting.RED));
             return 0;
         }
+
+        Feature feature = featureOptional.get();
 
         if (!feature.isEnabled()) {
             context.getSource()
