--- common/src/main/java/com/wynntils/core/features/DebugFeature.java
@@ -1,11 +1,0 @@
-/*
- * Copyright © Wynntils 2021.
- * This file is released under AGPLv3. See LICENSE for full license details.
- */
-package com.wynntils.core.features;
-
-import com.wynntils.core.features.properties.StartDisabled;
-
-/** Feature for debugging */
-@StartDisabled
-public abstract class DebugFeature extends UserFeature {}
